// src/src/generated/axios/api.ts
export interface DoctorDTO {
  /**
   *
   * @type {number}
   * @memberof DoctorDTO
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  surname?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  phoneNumber?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  email?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  avatar?: string;
  /**
   *
   * @type {string}
   * @memberof DoctorDTO
   */
  profession?: string;
  /**
   *
   * @type {Array<PatientDTO>}
   * @memberof DoctorDTO
   */
  latestPatients?: Array<PatientDTO>;
}
// ...
export interface PatientDTO {
  /**
   *
   * @type {number}
   * @memberof PatientDTO
   */
  id?: number;
  /**
   *
   * @type {number}
   * @memberof PatientDTO
   */
  opd?: number;
  /**
   *
   * @type {number}
   * @memberof PatientDTO
   */
  idp?: number;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  surname?: string;
  /**
   *
   * @type {number}
   * @memberof PatientDTO
   */
  phoneNumber?: number;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  email?: string;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  avatar?: string;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  bloodGroup?: PatientDTOBloodGroupEnum;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  notes?: string;
  /**
   *
   * @type {boolean}
   * @memberof PatientDTO
   */
  chronicPatient?: boolean;
  /**
   *
   * @type {string}
   * @memberof PatientDTO
   */
  lastAdmission?: string;
  /**
   *
   * @type {number}
   * @memberof PatientDTO
   */
  lastDoctorVisitedId?: number;
  /**
   *
   * @type {Array<PatientRecordDTO>}
   * @memberof PatientDTO
   */
  patientRecords?: Array<PatientRecordDTO>;
  /**
   *
   * @type {Array<number>}
   * @memberof PatientDTO
   */
  doctorIds?: Array<number>;
}

export const PatientDTOBloodGroupEnum = {
  ZeroPlus: "ZERO_PLUS",
  ZeroMinus: "ZERO_MINUS",
  APlus: "A_PLUS",
  AMinus: "A_MINUS",
  BPlus: "B_PLUS",
  BMinus: "B_MINUS",
  AbPlus: "AB_PLUS",
  AbMinus: "AB_MINUS",
} as const;

export type PatientDTOBloodGroupEnum = (typeof PatientDTOBloodGroupEnum)[keyof typeof PatientDTOBloodGroupEnum];

/**
 *
 * @export
 * @interface PatientFilterDTO
 */
export interface PatientFilterDTO {
  /**
   *
   * @type {string}
   * @memberof PatientFilterDTO
   */
  text?: string;
  /**
   *
   * @type {number}
   * @memberof PatientFilterDTO
   */
  id?: number;
  /**
   *
   * @type {number}
   * @memberof PatientFilterDTO
   */
  opd?: number;
  /**
   *
   * @type {number}
   * @memberof PatientFilterDTO
   */
  idp?: number;
  /**
   *
   * @type {number}
   * @memberof PatientFilterDTO
   */
  doctorId?: number;
}
/**
 *
 * @export
 * @interface PatientRecordDTO
 */
export interface PatientRecordDTO {
  /**
   *
   * @type {number}
   * @memberof PatientRecordDTO
   */
  id?: number;
  /**
   *
   * @type {number}
   * @memberof PatientRecordDTO
   */
  patientId?: number;
  /**
   *
   * @type {DoctorDTO}
   * @memberof PatientRecordDTO
   */
  doctor?: DoctorDTO;
  /**
   *
   * @type {string}
   * @memberof PatientRecordDTO
   */
  date?: string;
  /**
   *
   * @type {string}
   * @memberof PatientRecordDTO
   */
  typeVisit?: string;
  /**
   *
   * @type {string}
   * @memberof PatientRecordDTO
   */
  reasonVisit?: string;
  /**
   *
   * @type {string}
   * @memberof PatientRecordDTO
   */
  treatmentMade?: string;
}